refactor: improve cli_test.go organization by splitting into focused test files
- Split the large cli_test.go file (1007 lines) into multiple focused test files
- Created cli_args_test.go for command-line argument parsing tests
- Created cli_logging_test.go for logging configuration tests
- Created cli_validation_test.go for input validation tests
- Retained existing cli_token_test.go for token management tests
- Modified original file to be a stub that references new test files
- Marked task as complete in TODO.md
