+++ internal/integration/user_interaction_test.go
